requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-decouple==3.4
pytz==2021.1
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
whitenoise==5.2.0
zope.interface==5.3.0
